update design in sale report
